Impl door interaction animation + PiranhaFlower logic + code refactoring + bug fixes

+++ Assets/Scripts/Pal3/Scene/SceneObjects/StaticOrAnimatedObject.cs
@@ -69,7 +69,7 @@ namespace Pal3.Scene.SceneObjects
             // Parameters[2] describes animated object's default animation.
             // 0 means no animation. 1 means the object is animated up and down (sine curve).
             // 2 means the object is animated with constant rotation.
-            if ((int)_sceneObject.Info.Parameters[2] == 1)
+            if (_sceneObject.Info.Parameters[2] == 1)
             {
                 Transform currentTransform = transform;
                 Vector3 currentPosition = currentTransform.localPosition;
@@ -77,7 +77,7 @@ namespace Pal3.Scene.SceneObjects
                     _startYPosition + Mathf.Sin(Time.time) / 6f,
                     currentPosition.z);
             }
-            else if ((int)_sceneObject.Info.Parameters[2] == 2)
+            else if (_sceneObject.Info.Parameters[2] == 2)
             {
                 Transform currentTransform = transform;
                 transform.RotateAround(currentTransform.position, currentTransform.up, Time.deltaTime * 80f);
